Tempo for odd time signatures

+++ src/main/java/org/fergonco/music/mjargon/model/AbstractSongLine.java
@@ -20,7 +20,7 @@ public abstract class AbstractSongLine implements SongLine {
 	}
 
 	@Override
-	public int getTempo() {
+	public double getTempo() {
 		return 0;
 	}
