+++ src/main/java/org/fergonco/music/mjargon/model/SongLine.java
@@ -10,7 +10,7 @@ public interface SongLine {
 
 	boolean isTempo();
 
-	int getTempo();
+	double getTempo();
 
 	boolean isRepeat();
